rename dependencyAnalyzer to swaggerAnalyzer
