+++ lib/commands/analyze-dependency.ts
@@ -6,7 +6,7 @@ import * as yargs from "yargs";
 
 import { cliSuppressExceptions } from "../cliSuppressExceptions";
 import { getAutorestConfig } from "../util/getAutorestConfig";
-import { SwaggerAnalyzer } from "./../testScenario/dependencyAnalyzer";
+import { SwaggerAnalyzer } from "../testScenario/swaggerAnalyzer";
 
 export const command = "analyze-dependency";
